Move all classes to org.openhab.binding.elkm1.internal or subpackages
