+++ addons/binding/org.openhab.binding.elkm1/src/main/java/org/openhab/binding/elkm1/internal/ElkM1BindingConstants.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.elkm1;
+package org.openhab.binding.elkm1.internal;
 
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
