+++ addons/binding/org.openhab.binding.elkm1/src/main/java/org/openhab/binding/elkm1/internal/ElkM1HandlerFactory.java
@@ -20,11 +20,10 @@ import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
-import org.openhab.binding.elkm1.ElkM1BindingConstants;
-import org.openhab.binding.elkm1.discovery.ElkM1DiscoveryHandler;
-import org.openhab.binding.elkm1.handler.ElkM1AreaHandler;
-import org.openhab.binding.elkm1.handler.ElkM1BridgeHandler;
-import org.openhab.binding.elkm1.handler.ElkM1ZoneHandler;
+import org.openhab.binding.elkm1.internal.discovery.ElkM1DiscoveryHandler;
+import org.openhab.binding.elkm1.internal.handler.ElkM1AreaHandler;
+import org.openhab.binding.elkm1.internal.handler.ElkM1BridgeHandler;
+import org.openhab.binding.elkm1.internal.handler.ElkM1ZoneHandler;
 import org.osgi.framework.ServiceRegistration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
