+++ addons/binding/org.openhab.binding.elkm1/src/main/java/org/openhab/binding/elkm1/internal/discovery/ElkM1DiscoveryHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.elkm1.discovery;
+package org.openhab.binding.elkm1.internal.discovery;
 
 import java.util.Map;
 
@@ -14,10 +14,10 @@ import org.eclipse.smarthome.config.discovery.AbstractDiscoveryService;
 import org.eclipse.smarthome.config.discovery.DiscoveryResult;
 import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
 import org.eclipse.smarthome.core.thing.ThingUID;
-import org.openhab.binding.elkm1.ElkM1BindingConstants;
-import org.openhab.binding.elkm1.handler.ElkM1BridgeHandler;
+import org.openhab.binding.elkm1.internal.ElkM1BindingConstants;
 import org.openhab.binding.elkm1.internal.ElkM1HandlerListener;
 import org.openhab.binding.elkm1.internal.elk.ElkTypeToRequest;
+import org.openhab.binding.elkm1.internal.handler.ElkM1BridgeHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
